function-source/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-secret-manager>=0.0.0